pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "heavenly-eyes"
version = "1.0.0"
description = "HeavenlyEyes — All-seeing OSINT reconnaissance CLI tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "HeavenlyEyes Team"}
]
keywords = ["osint", "reconnaissance", "intelligence", "security", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "python-whois>=0.9.4",
    "dnspython>=2.4.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
heavenlyeyes = "heavenlyeyes.cli:app"
heyes = "heavenlyeyes.cli:app"

[project.urls]
Homepage = "https://github.com/limxi/heavenly-eyes"
Repository = "https://github.com/limxi/heavenly-eyes"

[tool.setuptools.packages.find]
include = ["heavenlyeyes*"]
